[Fix] 수정 내역
1. 추가 스탭 스킨 데이터 추가
2. 고용하지 않은 직원의 스킨 변경시 팝업이 발생하도록 변경
3. 스탭 스킨에 따라 스탭 창에서 테두리가 변경되도록 수정

+++ Assets/Scripts/UI/ChallengeUI/UIChallenge.cs
@@ -142,7 +142,7 @@ public class UIChallenge : MobileUIView
         _uiWeeklyAlarm.SetActive(_uiWeekly.UpdateUI());
     }
 
-    public void StartCoinAnime(int money, Vector3 coinPos)
+    public void StartCoinAnime(long money, Vector3 coinPos)
     {
         float time = 0;
         int coinCnt = 5;
@@ -170,7 +170,7 @@ public class UIChallenge : MobileUIView
         }
     }
 
-    public void StartDiaAnime(int money, Vector3 coinPos)
+    public void StartDiaAnime(long money, Vector3 coinPos)
     {
         float time = 0;
         int coinCnt = 5;
